pyproject.toml
[tool.poetry]
name = "habittracker"
version = "0.1.0"
description = ""
authors = ["Nurlan-Grachyov <nurlan.grachyov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.8"
djangorestframework = "^3.16.0"
black = "^25.1.0"
pyflakes = "3.3.2"
pycodestyle = "2.13.0"
mccabe = "==0.7.0"
poetry-core = "^2.1.2"
flake8 = "^7.2.0"
drf-yasg = "^1.21.10"
python-dotenv = "^1.1.0"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
requests = "^2.32.3"
celery = "^5.5.1"
django-celery-beat = "^2.7.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
